refactor: move Main code to other file and star clean slate
